Still need to fix dynamic length stuff
It is dying on reconversion. Later in the file, it dies on an aligned string (when reconverting to test), yet it seems most everything is getting changed properly, not sure what is causing this? Possibly the throwaway caused by the read copying? Perhaps the read copying is copying from the wrong location?